+++ CHANGELOG.md
@@ -619,3 +619,4 @@ This is a fast forward to v1.6.3 of daviscook477's fork with a few additional ch
 * Don't show card colors in compendium that contain no cards to avoid crash (kiooeht)
 * Fix `potion list` command changing potion rng (kiooeht)
 * Fix `potion list` command not including potions from other characters (kiooeht)
+* Fix `potion` command not giving potions from other characters (kiooeht)